crypto-bot/requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
asyncpg==0.29.0
websockets==12.0
python-decouple==3.8
asyncio-throttle==1.0.2
prometheus-client==0.19.0
pydantic==2.5.3
structlog==23.2.0